+++ enum/ActionId.ts
@@ -1,0 +1,3 @@
+export enum ActionId {
+	CLOSE_CHAT_BUTTON= 'sfla_close_chat',
+}
